Refactored Menu component to Fixed and Bubble folders
