+++ lib/components/Editor/Menu/Fixed/EmojiOption.jsx
@@ -5,7 +5,7 @@ import { Smiley } from "neetoicons";
 import Dropdown from "components/Common/Dropdown";
 import MenuButton from "components/Common/MenuButton";
 
-import EmojiPickerMenu from "../Emoji/EmojiPicker/EmojiPickerMenu";
+import EmojiPickerMenu from "../../CustomExtensions/Emoji/EmojiPicker/EmojiPickerMenu";
 
 const EmojiOption = ({ editor, theme = "light" }) => {
   const dropdownRef = useRef();
@@ -19,11 +19,11 @@ const EmojiOption = ({ editor, theme = "light" }) => {
       ref={dropdownRef}
       customTarget={() => (
         <MenuButton
+          color={theme === "dark" && "white"}
           data-cy="neeto-editor-fixed-menu-emoji-option-button"
           icon={Smiley}
           iconActive={dropdownRef?.current?._tippy?.state?.isVisible}
           tooltipProps={{ content: "Emoji", position: "bottom", delay: [500] }}
-          color={theme === "dark" && "white"}
         />
       )}
     >
