+++ lib/components/Editor/Menu/Fixed/FontSizeOption.jsx
