+++ lib/components/Editor/Menu/Fixed/TextColorOption.jsx
