+++ lib/components/Editor/Menu/Fixed/index.jsx
@@ -15,10 +15,10 @@ import {
   renderOptionButton,
 } from "./utils";
 
-import Mentions from "../Mention";
-import Variables from "../Variable";
+import Mentions from "../../CustomExtensions/Mention";
+import Variables from "../../CustomExtensions/Variable";
 
-const FixedMenu = ({
+const Fixed = ({
   editor,
   options,
   mentions = [],
@@ -79,4 +79,4 @@ const FixedMenu = ({
   );
 };
 
-export default FixedMenu;
+export default Fixed;
